Small schema fixes (#97)
1. Remove `st` as required field on layer. This field was removed in
https://github.com/lottie/lottie-spec/pull/53
2. Remove min/max on some gradient scalar properties. Scalar properties
are objects and the values are in `k`, so min/max doesn't apply at that
level. We would need to reach deeper into the object to properly apply
limits.

These were discovered using full strict mode on ajv. We may want to turn
this on in our tests, however I was having some issues with cases where
the property definition and required were split between if/then/allof
etc.